Renaming tuples to tuple
