+++ src/core3d/mod.rs
@@ -1,5 +1,5 @@
 pub mod array_base;
 pub mod coordinates4;
 pub mod point;
-pub mod tuples;
+pub mod tuple;
 pub mod vector;
